cmd/vet: check for sync types being copied during function calls
Using a type containing a sync type directly
in a function call (whether as a receiver,
a param, or a return value) is an easy way
to accidentally copy a lock or other sync primitive.
Check for it.

The test as implemented does not provide 100%
coverage; see the discussion near the bottom of
testdata/copylock.go for shortcomings.

Fixes issue 6729.

R=adg, r, dsymonds
CC=golang-dev
https://codereview.appspot.com/23420043

Committer: Rob Pike <r@golang.org>